// web/app/lib/gen/piano_user_list/v1/piano_user_list_pb.ts
export enum PianoListKind {
  /**
   * @generated from enum value: PIANO_LIST_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: PIANO_LIST_KIND_WISHLIST = 1;
   */
  WISHLIST = 1,

  /**
   * @generated from enum value: PIANO_LIST_KIND_VISITED = 2;
   */
  VISITED = 2,

  /**
   * @generated from enum value: PIANO_LIST_KIND_FAVORITE = 3;
   */
  FAVORITE = 3,
}
// ...
export class AddPianoToUserListRequest extends Message<AddPianoToUserListRequest> {
  /**
   * "pianos/{piano_id}"
   *
   * @generated from field: string parent = 1;
   */
  parent = "";

  /**
   * @generated from field: piano_user_list.v1.PianoListKind list_kind = 2;
   */
  listKind = PianoListKind.UNSPECIFIED;

  constructor(data?: PartialMessage<AddPianoToUserListRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "piano_user_list.v1.AddPianoToUserListRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "parent", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "list_kind", kind: "enum", T: proto3.getEnumType(PianoListKind) },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): AddPianoToUserListRequest {
    return new AddPianoToUserListRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): AddPianoToUserListRequest {
    return new AddPianoToUserListRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): AddPianoToUserListRequest {
    return new AddPianoToUserListRequest().fromJsonString(jsonString, options);
  }

  static equals(a: AddPianoToUserListRequest | PlainMessage<AddPianoToUserListRequest> | undefined, b: AddPianoToUserListRequest | PlainMessage<AddPianoToUserListRequest> | undefined): boolean {
    return proto3.util.equals(AddPianoToUserListRequest, a, b);
  }
}
// ...
export class RemovePianoFromUserListRequest extends Message<RemovePianoFromUserListRequest> {
  /**
   * "pianos/{piano_id}"
   *
   * @generated from field: string parent = 1;
   */
  parent = "";

  /**
   * @generated from field: piano_user_list.v1.PianoListKind list_kind = 2;
   */
  listKind = PianoListKind.UNSPECIFIED;

  constructor(data?: PartialMessage<RemovePianoFromUserListRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "piano_user_list.v1.RemovePianoFromUserListRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "parent", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "list_kind", kind: "enum", T: proto3.getEnumType(PianoListKind) },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): RemovePianoFromUserListRequest {
    return new RemovePianoFromUserListRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): RemovePianoFromUserListRequest {
    return new RemovePianoFromUserListRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): RemovePianoFromUserListRequest {
    return new RemovePianoFromUserListRequest().fromJsonString(jsonString, options);
  }

  static equals(a: RemovePianoFromUserListRequest | PlainMessage<RemovePianoFromUserListRequest> | undefined, b: RemovePianoFromUserListRequest | PlainMessage<RemovePianoFromUserListRequest> | undefined): boolean {
    return proto3.util.equals(RemovePianoFromUserListRequest, a, b);
  }
}
// ...
export class GetMyPianoUserListsRequest extends Message<GetMyPianoUserListsRequest> {
  /**
   * "pianos/{piano_id}"
   *
   * @generated from field: string parent = 1;
   */
  parent = "";

  constructor(data?: PartialMessage<GetMyPianoUserListsRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "piano_user_list.v1.GetMyPianoUserListsRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "parent", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GetMyPianoUserListsRequest {
    return new GetMyPianoUserListsRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GetMyPianoUserListsRequest {
    return new GetMyPianoUserListsRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GetMyPianoUserListsRequest {
    return new GetMyPianoUserListsRequest().fromJsonString(jsonString, options);
  }

  static equals(a: GetMyPianoUserListsRequest | PlainMessage<GetMyPianoUserListsRequest> | undefined, b: GetMyPianoUserListsRequest | PlainMessage<GetMyPianoUserListsRequest> | undefined): boolean {
    return proto3.util.equals(GetMyPianoUserListsRequest, a, b);
  }
}